stockbot-claude/requirements.txt
yfinance==0.2.36
python-dotenv==1.0.1
langchain-community==0.0.19
matplotlib==3.8.2
scipy==1.12.0
pymongo==4.6.1
requests==2.31.0
pandas==2.2.0
python-dateutil==2.8.2 